Solve HackerRank 'Breaking the Records' Challenge
Challenge Description:
Given an array of scores for a season for a player,
Write a function to determine the number of times the player
breaks their records for most and least points scored during the season.
Return an array with the numbers of times they broke their records.
Index 0 is for breaking most points records,
and index 1 is for breaking least points records.