Solve HackerRank 'Grading Students' Challenge
The challenge is to complete the gradingStudents function
which accepts a list of students' grades and rounds each student's grade
according to these rules:

- If the difference between the grade
and the next multiple of 5 is less than 3,
round grade up to the next multiple of 5.

- If the value of grade is less than 38,
no rounding occurs as the result will still be a failing grade.

The function returns a list of rounded grades