Solve HackerRank 'Number Line Jumps' Challenge
You are choreographing a circus show with various animals.
For one act, you are given two kangaroos on a number line ready to jump
in the positive direction (i.e, toward positive infinity).
You have to figure out a way to get both kangaroos at the same location
at the same time as part of the show.
If it is possible, return YES, otherwise return NO.