Solve HackerRank 'Smart Number 2' Challenge
In this challenge, the task is to debug the existing code to
successfully execute all provided test files.

A number is called a smart number if it has an odd number of factors.
Given some numbers, find whether they are smart numbers or not.

Debug the given function is_smart_number to correctly check
if a given number is a smart number.